# openapi/IDIS_OpenAPI_v6_3.yaml
openapi: 3.0.3
info:
  title: IDIS API (VC Edition) - v6.3
  version: "6.3"
  description: >
    Enterprise-grade API contract for IDIS (Institutional Deal Intelligence System), derived from IDIS v6.3.
    This API enables deal intake, document ingestion, claim registry access, Sanad chain retrieval,
    deterministic calculation runs, LangGraph debate runs, human verification gates, overrides,
    deliverable generation, audit logging, and integrations.
servers:
  - url: https://api.idis.example.com
    description: Production
  - url: https://api-staging.idis.example.com
    description: Staging
...
security:
  - bearerAuth: []
  - apiKeyAuth: []
paths:
...
  /v1/byok/key:
    post:
      tags: [ComplianceAdmin]
      summary: Configure the tenant BYOK key policy (ADMIN only; responses carry hash only)
      operationId: configureByokKey
      responses:
        "201":
          description: Key policy configured; the raw alias is never echoed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ByokKeyResponse"
        "400": { description: Invalid key alias }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { description: Forbidden }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: false
              required: [key_alias]
              properties:
                key_alias: { type: string, minLength: 1 }
  /v1/byok/key/rotate:
    post:
      tags: [ComplianceAdmin]
      summary: Rotate the tenant BYOK key to a new alias (ADMIN only)
      operationId: rotateByokKey
      responses:
        "200":
          description: Key rotated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ByokKeyResponse"
        "400": { description: Invalid key alias }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { description: Forbidden }
        "404": { $ref: "#/components/responses/NotFound" }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: false
              required: [key_alias]
              properties:
                key_alias: { type: string, minLength: 1 }
  /v1/byok/key/revoke:
    post:
      tags: [ComplianceAdmin]
      summary: Revoke the tenant BYOK key; Class2/3 access denies until re-keyed (ADMIN only)
      operationId: revokeByokKey
      responses:
        "200":
          description: Key revoked
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ByokKeyResponse"
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { description: Forbidden }
        "404": { $ref: "#/components/responses/NotFound" }
  /v1/legal-holds:
    post:
      tags: [ComplianceAdmin]
      summary: Apply a legal hold; the target cannot be deleted until lifted (ADMIN only)
      operationId: applyLegalHold
      responses:
        "201":
          description: Hold applied; the reason is hashed immediately and never echoed
          content:
            application/json:
              schema:
                type: object
                additionalProperties: false
                required: [hold_id, target_type, target_id, applied_at]
                properties:
                  hold_id: { type: string, format: uuid }
                  target_type: { type: string, enum: [DEAL, DOCUMENT, ARTIFACT] }
                  target_id: { type: string }
                  applied_at: { type: string, format: date-time }
        "400": { description: Invalid hold request }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { description: Forbidden }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: false
              required: [target_type, target_id, reason]
              properties:
                target_type: { type: string, enum: [DEAL, DOCUMENT, ARTIFACT] }
                target_id: { type: string, minLength: 1 }
                reason: { type: string }
  /v1/legal-holds/{holdId}/lift:
    post:
      tags: [ComplianceAdmin]
      summary: Lift a legal hold (ADMIN only; unknown or cross-tenant ids answer 404)
      operationId: liftLegalHold
      responses:
        "200":
          description: Hold lifted
          content:
            application/json:
              schema:
                type: object
                additionalProperties: false
                required: [hold_id, lifted_at]
                properties:
                  hold_id: { type: string, format: uuid }
                  lifted_at: { type: string, format: date-time }
        "400": { description: Hold already lifted }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { description: Forbidden }
        "404": { $ref: "#/components/responses/NotFound" }
  /v1/erasure-requests:
    post:
      tags: [Erasure]
      summary: Create a durable per-deal erasure request (ADMIN only; executed separately)
      operationId: createErasureRequest
      responses:
        "201":
          description: Erasure request created; the reason is hashed and never echoed
          content:
            application/json:
              schema:
                type: object
                additionalProperties: false
                required: [request_id, deal_id, status, requested_at]
                properties:
                  request_id: { type: string, format: uuid }
                  deal_id: { type: string, format: uuid }
                  status: { type: string, enum: [REQUESTED] }
                  requested_at: { type: string, format: date-time }
        "400": { description: Invalid erasure request }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { description: Forbidden }
        "404": { $ref: "#/components/responses/NotFound" }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: false
              required: [deal_id, reason]
              properties:
                deal_id: { type: string, format: uuid }
                reason: { type: string }
  /v1/erasure-requests/{requestId}/execute:
    post:
      tags: [Erasure]
      summary: Execute a pending erasure request (ADMIN only; hold-aware, full deal removal)
      operationId: executeErasureRequest
      responses:
        "200":
          description: Erasure executed; counts are safe-shape totals
          content:
            application/json:
              schema:
                type: object
                additionalProperties: false
                required: [request_id, status, executed_at, counts]
                properties:
                  request_id: { type: string, format: uuid }
                  status: { type: string, enum: [EXECUTED] }
                  executed_at: { type: string, format: date-time }
                  counts:
                    type: object
                    additionalProperties: { type: integer }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { description: Forbidden or blocked by legal hold }
        "404": { $ref: "#/components/responses/NotFound" }
        "409": { description: Already executed }
        "500": { description: Execution failed (request marked FAILED, re-executable) }
  /v1/compliance-exports:
    post:
      tags: [Erasure]
      summary: Build the tenant's sanitized compliance export bundle (ADMIN only)
      operationId: createComplianceExport
      responses:
        "201":
          description: Export bundle written; manifest carries metadata and hashes only
          content:
            application/json:
              schema:
                type: object
                additionalProperties: false
                required: [export_id, object_key, manifest_sha256, counts]
                properties:
                  export_id: { type: string, format: uuid }
                  object_key: { type: string }
                  manifest_sha256: { type: string }
                  counts:
                    type: object
                    additionalProperties: { type: integer }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { description: Forbidden }
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    apiKeyAuth:
      type: apiKey
      in: header
      name: X-IDIS-API-Key
...
  responses:
    Unauthorized:
      description: Unauthorized
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
    NotFound:
      description: Not Found
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
...
  schemas:
    Error:
      type: object
      required: [code, message]
      properties:
        code: { type: string, example: "bad_request" }
        message: { type: string, example: "Invalid field: deal.name" }
        details:
          type: object
          additionalProperties: true
        request_id: { type: string, example: "req_123" }
      example:
        code: "not_found"
        message: "Deal a1b2c3d4-e5f6-7890-abcd-ef1234567890 not found"
        details:
          resource_type: "deal"
          resource_id: "a1b2c3d4-e5f6-7890-abcd-ef1234567890"
        request_id: "req_8f3a2b1c-0000-0000-0000-000000000001"
...
    ByokKeyResponse:
      type: object
      additionalProperties: false
      required: [key_alias_hash, key_state]
      properties:
        key_alias_hash:
          type: string
          description: 16-hex SHA-256 prefix of the key alias; the raw alias is never returned
        key_state: { type: string, enum: [ACTIVE, REVOKED] }
...